+++ server/src/main/java/com/example/server/service/SysService.java
@@ -1,8 +1,6 @@
 package com.example.server.service;
 
-import com.example.server.entity.UserModel;
 import com.example.server.mapper.SysMapper;
-import com.example.server.mapper.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,11 +9,11 @@ public class SysService {
     @Autowired
     public SysMapper sysMapper;
 
-    public int saveUuid(int uuid) {
+    public long saveUuid(long uuid) {
         return sysMapper.saveUuid(uuid);
     }
 
-    public int getUuid() {
+    public long getUuid() {
         return sysMapper.getUuid();
     }
 }
